// app/components/types.ts
export type ButtonValue = 
  | '0' | '1' | '2' | '3' | '4' | '5' | '6' | '7' | '8' | '9'
  | '.' | 'C' | '±' | '%' | '÷' | '×' | '-' | '+' | '='
  | 'sin' | 'cos' | 'tan' | '√' | 'log' | 'ln' | 'π' | '^' | '(' | ')' | '!' | 'e'
  | '⌫'; // Backspace button

export interface CalculatorState {
  currentValue: string
  previousValue: string
  operation: string
  overwrite: boolean
}